Changed size field type to long long

+++ src/agent/Disc.h
@@ -6,12 +6,12 @@ class Disc
 {
 public:
     Disc();
-    Disc(const std::string& _caption, const std::string& _deviceId, const std::string& _model,  const int& _partitions, const int& _size);
+    Disc(const std::string& _caption, const std::string& _deviceId, const std::string& _model,  const int& _partitions, const long long& _size);
 
 private:
     std::string caption;
     std::string deviceId;
     std::string model;
     int partitions;
-    int size;
+    long long size;
 };
